uwsgi: save some memory reordering wsgi_request struct
Continuing the saving memory new year resolution this time is
wsgi_request time.

With the help of pahole (from dwarves) by shuffling some fields
we get from:

    /* size: 872, cachelines: 14, members: 183 */
    /* sum members: 803, holes: 33, sum holes: 69 */

to:

    /* size: 808, cachelines: 13, members: 183 */
    /* sum members: 803, holes: 2, sum holes: 5 *

Which are 64 bytes saved. struct uwsgi_core saves 64
bytes for free too.

Please note these measurements are done on a 32bit machine, but
since the savings are found by moving smaller that 32 bit fields
it should be the similar on 64bit.