Implement lingering close for TCP sockets (FastCGI protocol only).
Similar to
  nginx as a server
  httpd as a server
  flup as a FastCGI server
  Open Market FastCGI programming library (server)
  etc. (not just HTTP or SCGI or FastCGI)

See new comments for considerations for an efficient implementation.

It is nice that this will eat all input (subject to timeout), which
helps when the application sends an error message while the client
is still writing the request body.  (The client gets to see the
error message.)  But typically the server-side of the socket isn't
willing to read arbitrary amounts of data, so that feature would
be lost when the client is still writing body for a while.