Tighten the guarantees of chain reloading
Improve the chain reload logic to guarantee that all workers are running
a version of the application that is recent at least as of the point in
time when the chain reload was triggered.  Previously the chain
reloading logic did not pay enough attention to workers which were
loading the application at the moment when the chain reload was
triggered, which could lead to surprising behaviour.