Add wrapper for common uwsgi_routing_translate usage
uwsgi_routing_translate is called a couple dozen times the same
way so add a handy wrapper that avoids c&p of code to extract
the subject from the uwsgi route. The goal is to make the code
prettier.

This makes the code just bit smaller (plugins shrinks too):

Before:

   text	   data	    bss	    dec	    hex	filename
1200437	  49300	  15012	1264749	 134c6d	uwsgi

After:

   text	   data	    bss	    dec	    hex	filename
1199743	  49300	  15012	1264055	 1349b7	uwsgi