Add some missing return values checking
These i think are the uncontroversial, for most of them we do
nothing with the error, we can revise this case by case if we want
to do proper error checking. A comment has been added so it's
easy to find them.
The only one adding error checking is doing in websockets init,
since it is done at startup i think it's fair to exit(1).