Extend internal uwsgi timer api for handling ns value
We add to struct_timer a long field to store a value in ns. This
does not replace the int field for storing values with seconds
resolution.
The facing api is not changed so plugins and configuration to add
timers still works only with second resolution values.
A new high resolution version (*_hr suffix) is provided for both
uwsgi_add_timer and event_queue_add_timer.