Fix running on musl libc
This is following on from the work in PR #522 which reached the point that
uwsgi compiled on musl libc (Alpine Linux) however it doesn't run because it
fails on pthread_mutexattr_setprotocol(&attr, PTHREAD_PRIO_INHERIT) which
returns ENOTSUP and the code treated any non-zero return value as fatal.

This commit handles this specific case while still considering any other
non-zero return value as an error the same as before.

With this change uwsgi now runs on Alpine Linux.