backend_broke flag in mod_proxy_uwsgi.c is always 1
In uwsgi_response() the variable backend_broke is
used at the very end to decide if the function
success. However the variable is initiated to 1 and,
again, set to 1 in the else if (rv != APR_SUCCESS)
clause.

I do not think this is intentional but more like a bug.

Backport of 8212290aeb3631e3e5e85b3bb3796ebc2f0f56d4