core: Add a --skip-atexit-teardown option to skip perl/python teardown
This is a general option meant to be used for any number of embedded
languages, but which is currently just used by python & perl.

As detailed at length in my issue #1384[1] the full destruction I
added for perl in 1dcdb72[2] can be *very* expensive when a child
dies, and in particular when all children die at once when the server
goes down. This was first introduced in uWSGI 2.0.9.

This is not a perl-specific issue as the issue notes, but rather a
generic issue with embedding any number of programming language
interpreters & compilers. Fully tearing them down can be very
expensive, but just calling C's exit function is cheap, and often
people running uWSGI at scale are very willing to make that trade-off.

This issue also impacts python embedding since uWSGI 1.1 (f4f4d5a) as
noted in a follow-up comment of mine to that issue.

So to fix this add a --skip-atexit-teardown option to allow skipping
DESTROY during perl destruction, __del__ during python destruction
etc. This can then be used for other languages in the future as we add
uwsgi_*_atexit handlers for them.

This emulates the hack of calling POSIX::_exit() in perl or
os._exit(0) in python at the very end of the atexit handler, but
allows uWSGI to do any cleanup it wanted to do on its own after
calling atexit.

1. https://github.com/unbit/uwsgi/issues/1384
2. https://github.com/unbit/uwsgi/commit/1dcdb72