Fix SystemError exception when async_args PyTuple is reused. #1408
In the case where the uwsgi_python_create_env_cheat env behavior is
used, the async_args PyTuple is reused on every request.

In some cases, the reference count on async_args may be > 1 at the end
of a request/response cycle. If PyTuple_SetItem is called with a
reference count > 1, a SystemError is raised.

When the cheat env behavior is used, the environ dictionary is also
reused between requests. This enables the elimination of redundant
PyTuple_SetItem calls by first checking to see if the environ dictionary
is already in the async_args tuple.