Write x509 DER to the uwsgi buffer
This will write the full x.509 DER into the buffer for use by clients
during runtime. This feature is intended to allow clients to handle
per-user ACL with the direct x.509 Certificate, without having to
configure the webserver to extract the right bits, which may or may not
be custom extensions.

One such example would be using and extracting the UPN SAN, or some
other exotic extension.