Add ssl-verify-depth flag to set the max Client CA chain length
When verifying Certificates, OpenSSL will build a chain of trust from a
root CA to the Certificate, via intermediary CA(s). The chain was
hardcoded to 1 previously, which means that all Client Certificates
would have to be issued directly from a root CA.

This adds a new flag, --ssl-verify-depth with a default of 1 (existing
behavior) that will allow N intermediary CAs when building the chain of
trust.