new plugin hook: early_post_jail
post_jail plugin hooks runs after all other hooks in post_jail,
so it was not possible to configure for example tuntap interface
in a post-jail exec/hook, since it was not yet available.

This patch adds a early_post_jail phase that plugins can attach to,
and moves initialization of tuntap interface to this phase.