Tiny validation on request_uri
As a continue of ticket https://github.com/unbit/uwsgi/pull/1687.

This PR does a simple "request_uri" check and responds with "400 Bad request" to the client if the first line of request contains more than one space character.

p.s.
Litterally I couldn't find a common approach of responding on bad requests with 400 code, so applied a small patch directly in the "http_parse" function. Hope it's ok.