plugins/python: fix compilation with gcc-8 -Werror=cast-function-type
gcc does not like that we are casting to a function with a different
signature. It's indeed a sane think to do, unfortunately the Python
C API has another opinion on that.
Use the same trick to cast to void * as proposed in Python itself:
https://github.com/python/cpython/pull/6749/files

While at it drop python specific version ifdefery around
send_to_spooler and spool because they should take varargs
without issues on python2 too.