Fix segmentation fault during worker shutdown. #1651
Some plugins (such as python) rely on plugin-level atexit callbacks to
be invoked before interpreter teardown (when system-level atexit
callbacks are invoked).

Prior to this change uWSGI was loading Python applications after
registering atexit functions to invoke plugin-level atexit callbacks.

The segfault could be triggered if loading a Python application that
contained a C extension that registered an atexit callback. Those atexit
callbacks would be triggered prior to full termination of the Python
program. As a result, Python would be operating on a C extension that
was already in a cleaned-up state.

This fix makes sure all uwsgi plugin atexit callbacks (which operate on
the Python level, for example) are run prior to atexit callbacks that
were scheduled through a C extension loading. This way, all extensions
will be in a consistent state at the time that uwsgi runs the Python
atexit callbacks.