workaround for fork-serve/emperor respawn
Immplemented a workaround/fix for begin able
to respawn adopted vassals when in emperor mode

First it introduces another message for the emperor
protocol, when emperor recives '40' it removes that
vassal from the eventloop hindering the emperor from
cursing the vassal later on.

To hinder adopted vassals from begin respawned as
normal the "acidental death" logic at 2309:core/emperor.c
now adds atters if the emperor uses fork-server.

Last but not least vassals now sends a '40' to the
emperor befor exiting when the exit_on_reload is set
to 1.