Fix crash when application cleanup handler reads request body.
If an application cleanup handler tries to access the request body, it
will access dangling freed memory and a dangling closed FILE descriptor.
This can cause a crash of the uWSGI server.

uWSGI calls cleanup handlers after close_and_free_request(), which
closes and frees the post_buffering FILE descriptor and buffers.
The pointer are not set to NULL, but this is not enough as
uwsgi_request_body_read() will still use one of the buffers.

This commit adds a flag to stop using the buffers after they've been
freed. The uwsgi_request_body_*() method will return empty content.