mules: run a mem usage collector in each mule when --memory-report is given
The idea was taken from the worker mem_collector thread that is used for
the --evil-reload-on-{as,rss} options.

This patch may not respect the ideas of --memory-report and --mem-collector-freq
properly. However, it for mules it seems sensible to collect memory usage in a
separate thread and the options seemed applicable.

Currently the collector thread will just set the new rss_size/vsz_size
members added to struct uwsgi_mule. In the future it may be interesting
to provide the reload-on-{as,rss} options for mules as well.