uwsgi_opt_set_int: use strtol() and warn on non-numeric data
This should not change the existing behavior of using 0 when
a non-numeric value is provided, but loudly logs:

    $ uwsgi test.ini --processes 2x
    [uWSGI] getting INI configuration from test.ini
    [WARNING] non-numeric value "2x" for option "processes" - using 2 !
    [WARNING] non-numeric value "true" for option "http-keepalive" - using 0 !
    *** Starting uWSGI 2.1-dev-ade7d170 (64bit) on [Sun Jun 23 18:36:29 2019] ***
    compiled with version: 8.3.0 on 23 June 2019 16:22:05
    ...

It might be saner to not even start with invalid options. However,
hypothetical setups with http-keepalive=off or http-keepalive=false may
then fail to start at all and people might get unahppy.