Add a XML2_CONFIG environment variable for cross-compiling
Currently, xml2-config is called out with no way to define a xml2-config path.
This causes uwsgi to use the host xml2-config which will cuase the xml2 library
and cflag directories to point to the host.

Add a check for the XML2_CONFIG environment variable, and if it exists, use
the resulting path instead.