Add a PCRE_CONFIG environment variable for cross-compiling
Currently, pcre-config is called out with no way to define a pcre-config path,
which causes uwsgi to use the host pcre-config, which will cause the pcre
library and cflag directories to point to the host.

Add a check for the PCRE_CONFIG environment variable, and if it exists,
use the resulting path instead.