Add PLUGIN_BASE_DIR variable
Currently, when cross-compiling, if the plugin_dir points to the target
directory, uwsgi will embed the full path during compiling. This whole path
results in uwsgi trying to load a full target path instead of /usr/lib/uwsgi
when running on the target.

For example with Buildroot, instead of using /usr/lib/uwsgi as the plugin_dir,
uwsgi would try to use /path/to/buildroot/output/target/usr/lib/uwsgi.

Creating a new PLUGIN_BASE_DIR variable and prefixing plugin_dir allows the
plugin to be installed to the appropriate directory but still have uwsgi load
the plugins from the correct folder when ran from on the cross-compiled target.