Add docker-based build
Having a simple way to build uwsgi locally in Docker with all
dependencies might be helpful.

This PR adds a Dockerfile (based on the Travis configuration)
and updates the Makefile to let you run `make docker-test`.

The first time will be slow as the image builds, but future runs
should be quick as the docker cache will be used.