yaml/libyaml: Fix keys ignored after sequence, support non-indented sequence entries
This patch re-purposes the in_uwsgi_section variable to track the depth
of the configuration. Previously, leaving a nested sequence would
stop the parsing process, ignoring any keys following the sequence.

Further, support non-indented sequences for which no
YAML_BLOCK_SEQUENCE_START_TOKEN is emitted. The entries are preceded by
the YAML_BLOCK_ENTRY_TOKEN, however, which can be used instead to
recognize these entries.

The following configuration is parsed sensibly after this patch when
compiled with libyaml:

    uwsgi:
      http-socket: ':4040'
      wsgi: tests.hello
      static-map:
        - /static/js=/server/static/js
        - /static/css=/server/static/css
      cache2: [
          "name=thumbnails,items=200",
          "name=videos,items=100"
      ]
      env:  # this is missing the SEQUENCE_START_TOKEN
      - ENVIRONMENT=prod
      - "DATABASE_URL=mysql://localhost"
      workers: 4
      thunder-lock: true

    # Verified with --show-config
    ./uwsgi -y ./config.yaml --show-config
    ...
    ;uWSGI instance configuration
    [uwsgi]
    yaml = ./config.yaml
    http-socket = :4040
    wsgi = tests.hello
    static-map = /static/js=/server/static/js
    static-map = /static/css=/server/static/css
    cache2 = name=thumbnails,items=200
    cache2 = name=videos,items=100
    env = ENVIRONMENT=prod
    env = DATABASE_URL=mysql://localhost
    workers = 4
    thunder-lock = true
    show-config = true
    ;end of configuration

Fixes #2097 (but *only* if libyaml is used).