plugins/cgi: adds dontresolve option
This option permit to call the simbolic link instead of the file the simbolic link points.
All the security check are still done as the simbolic path is passed at the end after all the checks are passed. This is useful if some cgi app are used for multiple function based on the name they are called by.

Signed-off-by: Ansuel Smith <ansuelsmth@gmail.com>