plugins/psgi: fix impossibility of determining the end of the chunked stream
At the end of the stream, the uwsgi_chunked_read function returns
len 0 and a pointer to a buffer filled with old data.
Based on the description: https://perldoc.perl.org/5.16.1/perlapi.html,
The newSVpv function in the case when the len is zero uses the strlen.
This makes it impossible to determine the end of the stream since the
last chunk returns endlessly.