core: introduce a harakiri graceful period mechanism
Usually when a harakiri timeout is triggered, uwsgi master fires an SIGKILL to
the worker immediately. There is no time for the app to perform some operations
on the worker like triggering a stacktrace dump or generates a notification.
This commit introduces a harakiri delay time to make it easy and possible. It is
up to plugins like psgi to set a harakiri_delayed_at timestamp to tell the
master if it wants to give some time for the plugin to perform some special
operations like dumping stacktraces before the worker is killed.œ

The new delay time is controlled by a new option: harakiri-delay. By default,
no harakiri delay is applied.