plugins/psgi: add a harakiri handler and a pre-harakiri command
This commits makes use of the new harakiri graceful period mechanism. If a user
specifies the psgi-exec-pre-harakiri option, which is a separate program, then
psgi will trigger the program before harakiri and sets the harakiri_delayed_at,
which stops master from killing the worker immediately and delayed it after
harakiri-delay time (default is 0s).

There two environment variables is set for the `psgi-exec-pre-harakiri` command:

	UWSGI_HARAKIRI_WORKER_PID
	UWSGI_HARAKIRI_REQUEST_URI

A config example:

	harakiri-delay = 3
	psgi-exec-pre-harakiri = /bin/send_harakiri_event_with_stacktrace