python: raise OSError on closed connections
Check the wsgi_req->async_closed flag when processing response
iterations, and raise an OSError is the flag is set.

This solves the case, where a long running (but idle) client connection
is closed by the client. Since no data is written to the socket, the
close event is not detected, and the backend handler keeps running.

One case of an idle long running connection is log file streaming: The
client reads the log contents, and keeps hanging around for new log
messages. If no new messages occur, nothing is written to the
connection. When the client disconnects, the request handler should be
stopped.