Replace all `os.system` usage with `subprocess.call`
This aims to fix intermittent build failures on some platforms (most
notably Debian Bullseye) which appear to be due to the use of
`os.system` from background threads. The exact nature of the issue
is unclear, however moving to `subprocess.call` appears to fix it.

Fixes https://github.com/unbit/uwsgi/issues/2447.

While there's more that could be done here to modernise these calls,
this aims to be a minimal change which introduces subprocess without
much intrusion.